fix: changed profilenavigation to use the backend.
