+++ src/app/home/page.js
@@ -21,40 +21,6 @@ export default function Page() {
   );
 }
 
-function HomeUserData() {
-  const { data: session, status } = useSession();
-
-  // console.log(`User ID = ${session.user.name}`);
-
-  const sendID = { id: session.user.name };
-  /* Sends the ID of the user to backend. */
-
-  const fetcher = (...args) =>
-    fetch(
-      ...args,
-      {
-        method: "POST",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        
-        },
-        body: JSON.stringify(sendID),
-      },
-      {revalidateIfStale: false}
-      
-    ).then((res) => res.json());
-
-  const { data, error } = useSWR("/api/userById", fetcher);
-  /* Uses the SWR Next hook.  */
-
-  if (error || !data) {
-    return null;
-  }
-  console.log(data);
-
-  return data;
-}
 
 function Home() {
   //initialise the router for conditional redirection
@@ -96,16 +62,6 @@ function NavBar({ data }) {
   const recipesCollections = useRef(null);
   const foodCollection = useRef(null);
   const foodRecommendation = useRef(null);
-
-  let userSurname = "Smith",
-    userGender = "F";
-
-  data = HomeUserData();
-
-  if (data) {
-    userSurname = data.surname;
-    userGender = data.gender;
-  }
 
   const refList = [
     home,
@@ -154,7 +110,7 @@ function NavBar({ data }) {
             />
             <span>Basket</span>
           </div>
-          <ProfileNavigation name={"John, Doe"} gender={"M"} />
+          <ProfileNavigation/>
         </div>
       </div>
       <div className="bg-primary flex justify-center items-center gap-1 sm:overflow-hidden sm:flex-wrap sm:p-5">
